backend/requirements.txt
gunicorn
Django
djangorestframework
django-filter
django-url-filter
django-cors-headers
mysqlclient
humanize
coverage
coveralls
pep8
pyflakes
pytest
faker
pandas
xmltodict